+++ CHANGELOG.md
@@ -2,6 +2,18 @@ Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
+## 4.2.0-alpha.3
+
+### Changed
+
+#### RFPBundle
+* The name for `/admin/api/requests` REST API resource was changed to `/admin/api/rfqs`.
+* The name for `/admin/api/requestproducts` REST API resource was changed to `/admin/api/rfqproducts`.
+* The name for `/admin/api/requestproductitems` REST API resource was changed to `/admin/api/rfqproductitems`.
+* The name for `/admin/api/requestadditionalnotes` REST API resource was changed to `/admin/api/rfqadditionalnotes`.
+* The name for `/admin/api/rfpcustomerstatuses` REST API resource was changed to `/admin/api/rfqcustomerstatuses`.
+* The name for `/admin/api/rfpinternalstatuses` REST API resource was changed to `/admin/api/rfqinternalstatuses`.
+
 ## 4.2.0-alpha.2 (2020-05-29)
 [Show detailed list of changes](incompatibilities-4-2-alpha-2.md)